+++ libs/env-extractor/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "env-extractor"
-version = "0.0.5"
+version = "0.0.6"
 edition = "2018"
 
 # Fields to publish on crates.io
